+++ src/modules/fs/mv.mjs
@@ -2,15 +2,16 @@ import { resolve } from 'path';
 
 import processParams from '../../utils/helpers/processParams.mjs';
 import doesFileExist from '../../utils/helpers/doesFileExist.mjs';
+import throwOperationFailed from '../../utils/helpers/throwOperationFailed.mjs';
 import writeText from '../../utils/helpers/writeText.mjs';
 import getMessage from '../../utils/helpers/getMessage.mjs';
-import throwOperationFailed from '../../utils/helpers/throwOperationFailed.mjs';
 import copyFile from '../../utils/helpers/copyFile.mjs';
+import removeFile from '../../utils/helpers/removeFile.mjs';
 
-import colors from '../../utils/constants/colors.mjs';
 import errors from '../../utils/constants/errors.mjs';
+import colors from '../../utils/constants/colors.mjs';
 
-const cp = async (filePaths) => {
+const mv = async (filePaths) => {
   const sourceFilePath = processParams(filePaths[0]);
   const targetDirPath = processParams(filePaths[1]);
 
@@ -27,13 +28,14 @@ const cp = async (filePaths) => {
     }
 
     const onEnd = () => {
-      writeText(getMessage('FILE_COPIED', sourceFilePath, targetFilePath), colors.GREEN);
+      writeText(getMessage('FILE_MOVED', sourceFilePath, targetFilePath), colors.GREEN);
     };
 
     await copyFile(sourceFilePath, targetFilePath, onEnd);
+    await removeFile(sourceFilePath);
   } else {
     return throwOperationFailed(errors.FILE_NOT_FOUND);
   }
 };
 
-export default cp;
+export default mv;
