+++ src/modules/routeCommands.mjs
@@ -7,6 +7,7 @@ import add from './fs/add.mjs';
 import rn from './fs/rn.mjs';
 import cp from './fs/cp.mjs';
 import rm from './fs/rm.mjs';
+import mv from './fs/mv.mjs';
 
 import commands from '../utils/constants/commands.mjs';
 import errors from '../utils/constants/errors.mjs';
@@ -44,6 +45,9 @@ const routeCommands = async (line) => {
       break;
     case commands.RM:
       await rm(args);
+      break;
+    case commands.MV:
+      await mv(args);
       break;
     default:
       new CustomError(errors.INVALID_INPUT).displayErrorMessage();
