+++ src/utils/helpers/copyFile.mjs
@@ -2,7 +2,7 @@ import readFile from '../../modules/streams/readFile.mjs';
 import writeFile from '../../modules/streams/writeFile.mjs';
 import showCurrentPath from './showCurrentPath.mjs';
 
-const copyFile = (sourceFilePath, targetFilePath, onEnd) => {
+const copyFile = async (sourceFilePath, targetFilePath, onEnd) => {
   const onData = (writeStream) => {
     const readStream = readFile(
       sourceFilePath,
